Refactoring for New Stat Comp
Moved useState variables (dictionary, keyboard, error , and accuracy) into the app parent component to be shared with the TextBox and Stats components.
Created a new useState variable phraseRunTime to measure how long it takes the user to type a phrase that is yet to be implemented.
Accuracy Calculation is was moved to the parent component.
Error and Accuracy is being displayed in the new Stats Component.