Added Leaderboard and Profile
Added new Leaderboard component that will display the user statistics between the top players
Added functionality to the profile page to get the latest profile records.
Created a StatcCard component to display the numbers nicely.
Fixed bug that was not allowing for the right value to be computed.
Created a updateProfile Function in the will update and store the new values in user profile when rendered.