Added Settings
Added a pop up window accessible in the profile page through the settings button.
Settings Popup has option to change the username of the account or delete the account altogether.
Fixed validFormatUsername not returning false when there was no input.
Created a remove function to handle the firebase removal of documents.
Fixed bug that was allowing user change username through sign in button.