Fixed Stats not loading
Created an async function to handle waiting for the stats data to finish loading.
Fixed a few syntax bugs that where giving errors:
-changing username to display name
-adding () to snapshot.exists()
Added default value to return if there is an error in loadLeaderboard funciton.
Added useEffect on both sign up and login pop ups to handle clearing data in the input fields.
Added a footer to link the github.
Added CSS to handle when the window gets to narrow and move the components to center.