Login Error Handling
Fixed how the login system handled error when user inputs invalid email and passords.
Fixed how the login popup  was displaying the errors.
Destrctured the validSignUp function to reuse pieces like valid email and valid passord in the LoginUserWithEmail function.
Change the p tag in login and signup popup to an h3 and gave it the same padding.
Fixed changing username not updating the profile card.